Restructure into models, services and tasks mods to simplify
