+++ src/main.rs
@@ -1,10 +1,12 @@
+//! [Themelio](https://themelio.org) is a public blockchain focused on security, performance, and long-term stability
+
 use structopt::StructOpt;
-use themelio_core::*;
+use themelio_core;
 
 fn main() {
     env_logger::Builder::from_env("THEMELIO_LOG")
         .parse_filters("themelio_core")
         .init();
     let opts = Config::from_args();
-    smol::block_on(run_main(opts))
+    smol::block_on(themelio_core::run_main(opts))
 }
